Change: Split Dashboard component into the dashboard and a view
For the main dashboard on the start page the loading functions are used
and the connected redux store is bypassed. Therefore we are actually
using two different components. While on it the Dashboard component got
refactored to use hooks completely and the last used HOCs for redux got
dropped.
